kpi/rule versioning fix

+++ knowage/src/it/eng/spagobi/kpi/KpiService.java
@@ -17,46 +17,6 @@
  */
 package it.eng.spagobi.kpi;
 
-import java.io.IOException;
-import java.text.MessageFormat;
-import java.text.SimpleDateFormat;
-import java.util.ArrayList;
-import java.util.Calendar;
-import java.util.Collections;
-import java.util.Comparator;
-import java.util.Date;
-import java.util.GregorianCalendar;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.Iterator;
-import java.util.List;
-import java.util.Map;
-import java.util.Map.Entry;
-import java.util.Set;
-
-import javax.script.ScriptEngine;
-import javax.script.ScriptEngineManager;
-import javax.servlet.http.HttpServletRequest;
-import javax.ws.rs.DELETE;
-import javax.ws.rs.GET;
-import javax.ws.rs.POST;
-import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
-import javax.ws.rs.QueryParam;
-import javax.ws.rs.core.Context;
-import javax.ws.rs.core.Response;
-
-import org.apache.log4j.Logger;
-import org.json.JSONArray;
-import org.json.JSONException;
-import org.json.JSONObject;
-import org.quartz.JobExecutionException;
-
-import com.fasterxml.jackson.core.JsonGenerationException;
-import com.fasterxml.jackson.databind.JsonMappingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.mongodb.util.JSON;
-
 import it.eng.spago.error.EMFErrorSeverity;
 import it.eng.spago.error.EMFInternalError;
 import it.eng.spago.error.EMFUserError;
@@ -101,13 +61,55 @@ import it.eng.spagobi.tools.dataset.constants.DataSetConstants;
 import it.eng.spagobi.tools.datasource.bo.IDataSource;
 import it.eng.spagobi.tools.scheduler.bo.Trigger;
 import it.eng.spagobi.tools.scheduler.dao.ISchedulerDAO;
+import it.eng.spagobi.utilities.StringUtils;
 import it.eng.spagobi.utilities.exceptions.SpagoBIException;
 import it.eng.spagobi.utilities.exceptions.SpagoBIServiceException;
 import it.eng.spagobi.utilities.rest.RestUtilities;
 
+import java.io.IOException;
+import java.text.MessageFormat;
+import java.text.SimpleDateFormat;
+import java.util.ArrayList;
+import java.util.Calendar;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.Comparator;
+import java.util.Date;
+import java.util.GregorianCalendar;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.Iterator;
+import java.util.List;
+import java.util.Map;
+import java.util.Map.Entry;
+import java.util.Set;
+
+import javax.script.ScriptEngine;
+import javax.script.ScriptEngineManager;
+import javax.servlet.http.HttpServletRequest;
+import javax.ws.rs.DELETE;
+import javax.ws.rs.GET;
+import javax.ws.rs.POST;
+import javax.ws.rs.Path;
+import javax.ws.rs.PathParam;
+import javax.ws.rs.QueryParam;
+import javax.ws.rs.core.Context;
+import javax.ws.rs.core.Response;
+
+import org.apache.log4j.Logger;
+import org.json.JSONArray;
+import org.json.JSONException;
+import org.json.JSONObject;
+import org.quartz.JobExecutionException;
+
+import com.fasterxml.jackson.core.JsonGenerationException;
+import com.fasterxml.jackson.databind.JsonMappingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.mongodb.util.JSON;
+
 /**
  * @authors Salvatore Lupo (Salvatore.Lupo@eng.it)
- *
+ * 
  */
 @Path("/1.0/kpi")
 @ManageAuthorization
@@ -280,7 +282,7 @@ public class KpiService {
 	 * Executes a given query over a given datasource (dataSourceId) limited by maxItem param. It uses existing backend to retrieve data and metadata, but the
 	 * resulting json is lightened in order to give back something like this: {"columns": [{"name": "column_1", "label": "order_id"},...], "rows": [{"column_1":
 	 * "1"},...]}
-	 *
+	 * 
 	 * @param req
 	 * @return
 	 * @throws EMFUserError
@@ -348,18 +350,27 @@ public class KpiService {
 			String obj = RestUtilities.readBody(req);
 			Rule rule = (Rule) JsonConverter.jsonToObject(obj, Rule.class);
 
+			// Checking if query executes
 			executeQuery(rule.getDataSourceId(), rule.getDefinition(), 1, rule.getPlaceholders(), getProfile(req));
 
+			// Checking if aliases used by rule are all usable
 			IKpiDAO kpiDao = getKpiDAO(req);
 			Map<String, List<String>> aliasErrorMap = kpiDao.aliasValidation(rule);
+			JSError jsError = new JSError();
 			if (!aliasErrorMap.isEmpty()) {
-				JSONArray errors = new JSONArray();
-				JSError jsError = new JSError();
 				for (Entry<String, List<String>> error : aliasErrorMap.entrySet()) {
 					jsError.addErrorKey(error.getKey(), new JSONArray(error.getValue()).toString().replaceAll("[\\[\\]]", ""));
 				}
+			}
+			// Checking if any removed measure is linked to a kpi (if so we cannot save this rule)
+			Collection<String> removedMeasures = checkConstrainsWithKpi(rule, req);
+			if (!removedMeasures.isEmpty()) {
+				jsError.addErrorKey("newKpi.rule.usedByKpi.save.error", StringUtils.join(removedMeasures, ", "));
+			}
+			if (jsError.hasErrors()) {
 				return Response.ok(jsError.toString()).build();
 			}
+
 			return Response.ok().build();
 		} catch (Exception e) {
 			throw new SpagoBIServiceException(req.getPathInfo(), e);
@@ -376,10 +387,19 @@ public class KpiService {
 			Rule rule = (Rule) JsonConverter.jsonToObject(requestVal, Rule.class);
 			Integer id = rule.getId();
 			Integer version = rule.getVersion();
+			JSError jsError = new JSError();
 			// Rule name must be unique
 			Integer otherId = dao.getRuleIdByName(rule.getName());
 			if (otherId != null && (id == null || !id.equals(otherId))) {
-				return Response.ok(new JSError().addErrorKey(NEW_KPI_RULE_NAME_NOT_AVAILABLE, rule.getName()).toString()).build();
+				jsError.addErrorKey(NEW_KPI_RULE_NAME_NOT_AVAILABLE, rule.getName());
+			}
+			// Checking if any removed measure is linked to a kpi (if so we cannot save this rule)
+			Collection<String> removedMeasures = checkConstrainsWithKpi(rule, req);
+			if (!removedMeasures.isEmpty()) {
+				jsError.addErrorKey("newKpi.rule.usedByKpi.delete.error", StringUtils.join(removedMeasures, ", "));
+			}
+			if (jsError.hasErrors()) {
+				return Response.ok(jsError.toString()).build();
 			}
 			if (id == null) {
 				// Save a new Rule
@@ -480,11 +500,7 @@ public class KpiService {
 			if (kpi.getId() == null) {
 				dao.insertKpi(kpi);
 			} else {
-				if (kpi.isEnableVersioning()) {
-					dao.insertNewVersionKpi(kpi);
-				} else {
-					dao.updateKpi(kpi);
-				}
+				dao.updateKpi(kpi);
 			}
 
 			return Response.ok().build();
@@ -497,12 +513,11 @@ public class KpiService {
 	@DELETE
 	@Path("/{id}/{version}/deleteKpi")
 	@UserConstraint(functionalities = { SpagoBIConstants.KPI_MANAGEMENT })
-	public Response deleteKpi(@PathParam("id") Integer id, @PathParam("version") Integer version, @QueryParam("toBeVersioned") Boolean toBeVersioned,
-			@Context HttpServletRequest req) throws EMFUserError {
+	public Response deleteKpi(@PathParam("id") Integer id, @PathParam("version") Integer version, @Context HttpServletRequest req) throws EMFUserError {
 		IKpiDAO dao = getKpiDAO(req);
 		JSError jsError = checkKpiRel(dao, id, version);
 		if (!jsError.hasErrors()) {
-			dao.removeKpi(id, version, Boolean.TRUE.equals(toBeVersioned));
+			dao.removeKpi(id, version);
 			return Response.ok().build();
 		} else {
 			return Response.ok(jsError.toString()).build();
@@ -515,6 +530,20 @@ public class KpiService {
 	public Response deleteRule(@PathParam("id") Integer id, @PathParam("version") Integer version, @Context HttpServletRequest req) throws EMFUserError {
 		// Rule can only be removed logically
 		IKpiDAO dao = getKpiDAO(req);
+		Map<Kpi, List<String>> kpimap = dao.listKpisLinkedToRule(id, version, true);
+		if (!kpimap.isEmpty()) {
+			StringBuilder kpiNames = new StringBuilder();
+			Iterator<Kpi> iKpiset = kpimap.keySet().iterator();
+			while (iKpiset.hasNext()) {
+				Kpi kpi = iKpiset.next();
+				kpiNames.append(kpi.getName());
+				if (iKpiset.hasNext()) {
+					kpiNames.append(" ,");
+				}
+			}
+			return Response.ok(new JSError().addErrorKey("newKpi.rule.usedByKpi.delete.error", kpiNames.toString()).toString()).build();
+		}
+		// Rule can only be removed logically
 		dao.removeRule(id, version, true);
 		return Response.ok().build();
 	}
@@ -663,101 +692,6 @@ public class KpiService {
 
 	}
 
-	private JSError checkKpiRel(IKpiDAO dao, Integer id, Integer version) {
-		JSError jsError = new JSError();
-		// Check if any relation exists with Scorecard
-		List<Scorecard> scorecards = dao.listScorecardByKpi(id, version);
-		if (scorecards != null && !scorecards.isEmpty()) {
-			String scNames = "";
-			for (int i = 0; i < scorecards.size(); i++) {
-				if (i != 0) {
-					scNames += ", ";
-				}
-				scNames += scorecards.get(i).getName();
-			}
-			jsError.addErrorKey("newKpi.kpi.kpiIsUsedByScorecards", scNames);
-		}
-		// Check if any relation exists with Target
-		List<Target> targets = dao.listTargetByKpi(id, version);
-		if (targets != null && !targets.isEmpty()) {
-			String names = "";
-			for (int i = 0; i < targets.size(); i++) {
-				if (i != 0) {
-					names += ", ";
-				}
-				names += targets.get(i).getName();
-			}
-			jsError.addErrorKey("newKpi.kpi.kpiIsUsedByTargets", names);
-		}
-		// Check if any relation exists with Scheduler
-		List<KpiScheduler> schedulers = dao.listSchedulerByKpi(id, version);
-		if (schedulers != null && !schedulers.isEmpty()) {
-			String names = "";
-			for (int i = 0; i < schedulers.size(); i++) {
-				if (i != 0) {
-					names += ", ";
-				}
-				names += schedulers.get(i).getName();
-			}
-			jsError.addErrorKey("newKpi.kpi.kpiIsUsedBySchedulers", names);
-		}
-		return jsError;
-	}
-
-	private JSError check(Target target, IKpiDAO dao) throws SpagoBIException {
-		JSError e = new JSError();
-		if (target.getName() == null) {
-			e.addError("Name id mandatory");
-		}
-		if (target.getStartValidity() == null) {
-			e.addError("StartValidity is mandatory");
-		}
-		if (target.getEndValidity() == null) {
-			e.addError("EndValidity is mandatory");
-		}
-		if (target.getStartValidity() != null && target.getEndValidity() != null && target.getStartValidity().after(target.getEndValidity())) {
-			e.addErrorKey("newKpi.target.invalidPeriod");
-		}
-		if (target.getValues() == null || target.getValues().isEmpty()) {
-			e.addError("Values are mandatory");
-		}
-		// start/end validity dates of targets with same kpis cannot overlap
-		Set<Kpi> kpis = new HashSet<>();
-		for (TargetValue targetValue : target.getValues()) {
-			kpis.add(new Kpi(targetValue.getKpiId(), targetValue.getKpiVersion()));
-		}
-		List<Target> ll = dao.listOverlappingTargets(target.getId(), target.getStartValidity(), target.getEndValidity(), kpis);
-		String names = "";
-		for (int i = 0; i < ll.size(); i++) {
-			if (i != 0) {
-				names += ", ";
-			}
-			Target sameDateTarget = ll.get(i);
-			names += sameDateTarget.getName();
-		}
-		if (!names.isEmpty()) {
-			e.addErrorKey("newKpi.target.alreadyExistingPeriod", names);
-		}
-		return e;
-	}
-
-	private void checkValidity(KpiScheduler scheduler) throws SpagoBIException {
-		String fieldValue = null;
-		String fieldName = null;
-		if (!scheduler.getFrequency().getStartTime().matches("\\d{2}:\\d{2}")) {
-			fieldName = "StartTime";
-			fieldValue = scheduler.getFrequency().getStartTime();
-		}
-		if (scheduler.getFrequency().getEndTime() != null && !scheduler.getFrequency().getEndTime().isEmpty()
-				&& !scheduler.getFrequency().getEndTime().matches("\\d{2}:\\d{2}")) {
-			fieldName = "EndTime";
-			fieldValue = scheduler.getFrequency().getEndTime();
-		}
-		if (fieldName != null) {
-			throw new SpagoBIException("Field error: " + fieldName + "[" + fieldValue + "]");
-		}
-	}
-
 	/*
 	 * Gets a criterion id and a list of ScorecardStatus and returns a status
 	 */
@@ -881,6 +815,125 @@ public class KpiService {
 	 * *** Private methods ***
 	 */
 
+	private Collection<String> checkConstrainsWithKpi(Rule rule, HttpServletRequest req) throws EMFUserError {
+		Collection<String> measureAndKpi = new HashSet<>();
+		IKpiDAO kpiDao = getKpiDAO(req);
+		Map<Kpi, List<String>> kpimap = kpiDao.listKpisLinkedToRule(rule.getId(), rule.getVersion(), true);
+		Set<String> usedMeasureList = new HashSet<>();
+		for (List<String> m : kpimap.values()) {
+			usedMeasureList.addAll(m);
+		}
+		Set<String> newMeasureList = new HashSet<>();
+		for (RuleOutput ro : rule.getRuleOutputs()) {
+			newMeasureList.add(ro.getAlias());
+		}
+		usedMeasureList.removeAll(newMeasureList);
+		for (String name : usedMeasureList) {
+			for (Entry<Kpi, List<String>> kpi : kpimap.entrySet()) {
+				if (kpi.getValue().contains(name)) {
+					measureAndKpi.add("'" + name + "' used by '" + kpi.getKey().getName() + "'");
+					break;
+				}
+			}
+		}
+		return measureAndKpi;
+	}
+
+	private JSError checkKpiRel(IKpiDAO dao, Integer id, Integer version) {
+		JSError jsError = new JSError();
+		// Check if any relation exists with Scorecard
+		List<Scorecard> scorecards = dao.listScorecardByKpi(id, version);
+		if (scorecards != null && !scorecards.isEmpty()) {
+			String scNames = "";
+			for (int i = 0; i < scorecards.size(); i++) {
+				if (i != 0) {
+					scNames += ", ";
+				}
+				scNames += scorecards.get(i).getName();
+			}
+			jsError.addErrorKey("newKpi.kpi.kpiIsUsedByScorecards", scNames);
+		}
+		// Check if any relation exists with Target
+		List<Target> targets = dao.listTargetByKpi(id, version);
+		if (targets != null && !targets.isEmpty()) {
+			String names = "";
+			for (int i = 0; i < targets.size(); i++) {
+				if (i != 0) {
+					names += ", ";
+				}
+				names += targets.get(i).getName();
+			}
+			jsError.addErrorKey("newKpi.kpi.kpiIsUsedByTargets", names);
+		}
+		// Check if any relation exists with Scheduler
+		List<KpiScheduler> schedulers = dao.listSchedulerByKpi(id, version);
+		if (schedulers != null && !schedulers.isEmpty()) {
+			String names = "";
+			for (int i = 0; i < schedulers.size(); i++) {
+				if (i != 0) {
+					names += ", ";
+				}
+				names += schedulers.get(i).getName();
+			}
+			jsError.addErrorKey("newKpi.kpi.kpiIsUsedBySchedulers", names);
+		}
+		return jsError;
+	}
+
+	private JSError check(Target target, IKpiDAO dao) throws SpagoBIException {
+		JSError e = new JSError();
+		if (target.getName() == null) {
+			e.addError("Name id mandatory");
+		}
+		if (target.getStartValidity() == null) {
+			e.addError("StartValidity is mandatory");
+		}
+		if (target.getEndValidity() == null) {
+			e.addError("EndValidity is mandatory");
+		}
+		if (target.getStartValidity() != null && target.getEndValidity() != null && target.getStartValidity().after(target.getEndValidity())) {
+			e.addErrorKey("newKpi.target.invalidPeriod");
+		}
+		if (target.getValues() == null || target.getValues().isEmpty()) {
+			e.addError("Values are mandatory");
+		}
+		// start/end validity dates of targets with same kpis cannot overlap
+		Set<Kpi> kpis = new HashSet<>();
+		for (TargetValue targetValue : target.getValues()) {
+			kpis.add(new Kpi(targetValue.getKpiId(), targetValue.getKpiVersion()));
+		}
+		List<Target> ll = dao.listOverlappingTargets(target.getId(), target.getStartValidity(), target.getEndValidity(), kpis);
+		String names = "";
+		for (int i = 0; i < ll.size(); i++) {
+			if (i != 0) {
+				names += ", ";
+			}
+			Target sameDateTarget = ll.get(i);
+			names += sameDateTarget.getName();
+		}
+		if (!names.isEmpty()) {
+			e.addErrorKey("newKpi.target.alreadyExistingPeriod", names);
+		}
+		return e;
+	}
+
+	private void checkValidity(KpiScheduler scheduler) throws SpagoBIException {
+		String fieldValue = null;
+		String fieldName = null;
+		if (!scheduler.getFrequency().getStartTime().matches("\\d{2}:\\d{2}")) {
+			fieldName = "StartTime";
+			fieldValue = scheduler.getFrequency().getStartTime();
+		}
+		if (scheduler.getFrequency().getEndTime() != null && !scheduler.getFrequency().getEndTime().isEmpty()
+				&& !scheduler.getFrequency().getEndTime().matches("\\d{2}:\\d{2}")) {
+			fieldName = "EndTime";
+			fieldValue = scheduler.getFrequency().getEndTime();
+		}
+		if (fieldName != null) {
+			throw new SpagoBIException("Field error: " + fieldName + "[" + fieldValue + "]");
+		}
+	}
+
 	private void checkMandatory(KpiScheduler scheduler) throws SpagoBIException {
 		String fieldName = null;
 		if (scheduler.getName() == null) {
@@ -920,7 +973,7 @@ public class KpiService {
 
 	/**
 	 * Check if placeholders with default value are a subset of placeholders linked to measures used in kpi definition (ie kpi formula)
-	 *
+	 * 
 	 * @param servlet
 	 *            request
 	 * @param placeholder
